Add a menu to the ArtifactMapPane to copy out its JSON
